Cache descriptor defaults
