+++ build.gradle.kts
@@ -4,7 +4,7 @@ plugins {
 
 allprojects {
     group = "space.kscience"
-    version = "0.4.0"
+    version = "0.4.1"
 }
 
 subprojects {
