+++ dataforge-meta/src/commonMain/kotlin/space/kscience/dataforge/meta/descriptors/NodeDescriptor.kt
@@ -6,13 +6,29 @@ import space.kscience.dataforge.names.*
 
 
 /**
+ * A [Meta] that is constructed from [NodeDescriptor]
+ */
+private class DescriptorMeta(val descriptor: NodeDescriptor) : Meta, MetaBase() {
+    override val items: Map<NameToken, MetaItem>
+        get() = buildMap {
+            descriptor.items.forEach { (token, descriptorItem) ->
+                val item = descriptorItem.defaultValue
+                if (item != null) {
+                    put(NameToken(token), item)
+                }
+            }
+        }
+}
+
+
+/**
  * Descriptor for meta node. Could contain additional information for viewing
  * and editing.
  *
  * @author Alexander Nozik
  */
 @DFBuilder
-public sealed interface NodeDescriptor: ItemDescriptor {
+public sealed interface NodeDescriptor : ItemDescriptor {
     /**
      * True if the node is required
      *
@@ -41,6 +57,11 @@ public sealed interface NodeDescriptor: ItemDescriptor {
      * The list of children value descriptors
      */
     public val values: Map<String, ValueDescriptor>
+
+    /**
+     * Generate a laminate representing default item set generated by this descriptor
+     */
+    public fun defaultMeta(): Laminate = Laminate(default, DescriptorMeta(this))
 
     public companion object {
 
@@ -165,6 +186,11 @@ public class NodeDescriptorBuilder(config: Config = Config()) : ItemDescriptorBu
     public fun value(name: String, block: ValueDescriptorBuilder.() -> Unit) {
         value(name.toName(), block)
     }
+
+    /**
+     * Build a default [MetaItemNode] from this node descriptor
+     */
+    override val defaultValue: MetaItem by lazy { MetaItemNode(defaultMeta()) }
 
     override fun build(): NodeDescriptor = NodeDescriptorBuilder(config.copy())
